Replace CryptoJS with Web Crypto AES-GCM
The original motivation for this was just that CryptoJS isn't
authenticated, meaning we can't reliably tell if the decryption was
actually successful.

Modern browser support is good, although note that the api does require
a secure context (HTTPS) to be used.

A legacy path is kept for decryption to let any existing CryptoJS
pastes keep working.