Change themes again
Swapped the dark theme as it had a number of shortcomings, then rotated
some colors in the light theme to be more consistent with it. Then went
through the dark theme and drastically muted the colors.

Still not a big fan of either theme but I don't have more energy to put
in. I am aware the dark theme is uncomfortably low contrast now...