Rename button "Encrypt", combine user password with uri password
Previously the button was less intuitively called "Random" [password]
and disabled the password field when active.

Previously when the user supplied a password we just used it as is.
Trivial passwords are trivially brute forced in a stolen database, so
it's useful to also store secure material in the uri fragment. This way
decryption also requires possession of the full url.