Mark more async and run db put/get as blocking
This is a late part 2 of the rocket 0.5.0 commit, since everything is
already run in async runtime now and this was demanded in the porting
guide.

Notes:

The db put/get is probably not really heavy enough to need this wrapper.
Maybe it could/should have been left alone...

The other small refactor around DBWithThreadMode<SingleThreaded> is
because I worried we should use <MultiThreaded>, but looking closer
decided it's not needed. The docs suggest <MultiThreaded> is only to
"allow column families to be created and dropped from multiple threads
concurrently".